Struct rusoto_gamelift::Build

pub struct Build {
    pub build_id: Option<String>,
    pub creation_time: Option<f64>,
    pub name: Option<String>,
    pub operating_system: Option<String>,
    pub size_on_disk: Option<i64>,
    pub status: Option<String>,
    pub version: Option<String>,
}

Properties describing a game build.

Build-related operations include:

CreateBuild

ListBuilds

DescribeBuild

UpdateBuild

DeleteBuild

Fields
build_id: Option<String>

Unique identifier for a build.

creation_time: Option<f64>

Time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

name: Option<String>

Descriptive label that is associated with a build. Build names do not need to be unique. It can be set using CreateBuild or UpdateBuild.

operating_system: Option<String>

Operating system that the game server binaries are built to run on. This value determines the type of fleet resources that you can use for this build.

size_on_disk: Option<i64>

File size of the uploaded game build, expressed in bytes. When the build status is INITIALIZED, this value is 0.

status: Option<String>

Current status of the build.

Possible build statuses include the following:

INITIALIZED – A new build has been defined, but no files have been uploaded. You cannot create fleets for builds that are in this status. When a build is successfully created, the build status is set to this value.

READY – The game build has been successfully uploaded. You can now create new fleets for this build.

FAILED – The game build upload failed. You cannot create new fleets for this build.

version: Option<String>

Version that is associated with this build. Version strings do not need to be unique. This value can be set using CreateBuild or UpdateBuild.