Struct rusoto_gamelift::CreateFleetInput

pub struct CreateFleetInput {
    pub build_id: String,
    pub description: Option<String>,
    pub ec2_inbound_permissions: Option<Vec<IpPermission>>,
    pub ec2_instance_type: String,
    pub log_paths: Option<Vec<String>>,
    pub metric_groups: Option<Vec<String>>,
    pub name: String,
    pub new_game_session_protection_policy: Option<String>,
    pub resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>,
    pub runtime_configuration: Option<RuntimeConfiguration>,
    pub server_launch_parameters: Option<String>,
    pub server_launch_path: Option<String>,
}

Represents the input for a request action.

Fields
build_id: String

Unique identifier for a build to be deployed on the new fleet. The build must have been successfully uploaded to Amazon GameLift and be in a READY status. This fleet setting cannot be changed once the fleet is created.

description: Option<String>

Human-readable description of a fleet.

ec2_inbound_permissions: Option<Vec<IpPermission>>

Range of IP addresses and port settings that permit inbound traffic to access server processes running on the fleet. If no inbound permissions are set, including both IP address range and port range, the server processes in the fleet cannot accept connections. You can specify one or more sets of permissions for a fleet.

ec2_instance_type: String

Name of an EC2 instance type that is supported in Amazon GameLift. A fleet instance type determines the computing resources of each instance in the fleet, including CPU, memory, storage, and networking capacity. Amazon GameLift supports the following EC2 instance types. See Amazon EC2 Instance Types for detailed descriptions.

log_paths: Option<Vec<String>>

This parameter is no longer used. Instead, to specify where Amazon GameLift should store log files once a server process shuts down, use the Amazon GameLift server API ProcessReady() and specify one or more directory paths in logParameters. See more information in the Server API Reference.

metric_groups: Option<Vec<String>>

Names of metric groups to add this fleet to. Use an existing metric group name to add this fleet to the group, or use a new name to create a new metric group. Currently, a fleet can only be included in one metric group at a time.

name: String

Descriptive label that is associated with a fleet. Fleet names do not need to be unique.

new_game_session_protection_policy: Option<String>

Game session protection policy to apply to all instances in this fleet. If this parameter is not set, instances in this fleet default to no protection. You can change a fleet's protection policy using UpdateFleetAttributes, but this change will only affect sessions created after the policy change. You can also set protection for individual instances using UpdateGameSession.

NoProtection – The game session can be terminated during a scale-down event.

FullProtection – If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>

Policy that limits the number of game sessions an individual player can create over a span of time for this fleet.

runtime_configuration: Option<RuntimeConfiguration>

Instructions for launching server processes on each instance in the fleet. The runtime configuration for a fleet has a collection of server process configurations, one for each type of server process to run on an instance. A server process configuration specifies the location of the server executable, launch parameters, and the number of concurrent processes with that configuration to maintain on each instance. A CreateFleet request must include a runtime configuration with at least one server process configuration; otherwise the request will fail with an invalid request exception. (This parameter replaces the parameters ServerLaunchPath and ServerLaunchParameters; requests that contain values for these parameters instead of a runtime configuration will continue to work.)

server_launch_parameters: Option<String>

This parameter is no longer used. Instead, specify server launch parameters in the RuntimeConfiguration parameter. (Requests that specify a server launch path and launch parameters instead of a runtime configuration will continue to work.)

server_launch_path: Option<String>

This parameter is no longer used. Instead, specify a server launch path using the RuntimeConfiguration parameter. (Requests that specify a server launch path and launch parameters instead of a runtime configuration will continue to work.)