Struct rusoto_gamelift::DescribeGameSessionsInput

pub struct DescribeGameSessionsInput {
    pub alias_id: Option<String>,
    pub fleet_id: Option<String>,
    pub game_session_id: Option<String>,
    pub limit: Option<i64>,
    pub next_token: Option<String>,
    pub status_filter: Option<String>,
}

Represents the input for a request action.

Fields
alias_id: Option<String>

Unique identifier for an alias associated with the fleet to retrieve all game sessions for.

fleet_id: Option<String>

Unique identifier for a fleet to retrieve all game sessions for.

game_session_id: Option<String>

Unique identifier for the game session to retrieve. You can use either a GameSessionId or GameSessionArn value.

limit: Option<i64>

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

next_token: Option<String>

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To specify the start of the result set, do not specify a value.

status_filter: Option<String>

Game session status to filter results on. Possible game session statuses include ACTIVE, TERMINATED, ACTIVATING, and TERMINATING (the last two are transitory).