Struct rusoto_gamelift::FleetAttributes

pub struct FleetAttributes {
    pub build_id: Option<String>,
    pub creation_time: Option<f64>,
    pub description: Option<String>,
    pub fleet_arn: Option<String>,
    pub fleet_id: Option<String>,
    pub log_paths: Option<Vec<String>>,
    pub metric_groups: Option<Vec<String>>,
    pub name: Option<String>,
    pub new_game_session_protection_policy: Option<String>,
    pub operating_system: Option<String>,
    pub resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>,
    pub server_launch_parameters: Option<String>,
    pub server_launch_path: Option<String>,
    pub status: Option<String>,
    pub termination_time: Option<f64>,
}

General properties describing a fleet.

Fields
build_id: Option<String>

Unique identifier for a build.

creation_time: Option<f64>

Time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

description: Option<String>

Human-readable description of the fleet.

fleet_arn: Option<String>

Identifier for a fleet that is unique across all regions.

fleet_id: Option<String>

Unique identifier for a fleet.

log_paths: Option<Vec<String>>

Location of default log files. When a server process is shut down, Amazon GameLift captures and stores any log files in this location. These logs are in addition to game session logs; see more on game session logs in the Amazon GameLift Developer Guide. If no default log path for a fleet is specified, Amazon GameLift will automatically upload logs that are stored on each instance at C:\game\logs (for Windows) or /local/game/logs (for Linux). Use the Amazon GameLift console to access stored logs.

metric_groups: Option<Vec<String>>

Names of metric groups that this fleet is included in. In Amazon CloudWatch, you can view metrics for an individual fleet or aggregated metrics for a fleets that are in a fleet metric group. Currently, a fleet can be included in only one metric group at a time.

name: Option<String>

Descriptive label that is associated with a fleet. Fleet names do not need to be unique.

new_game_session_protection_policy: Option<String>

Type of game session protection to set for all new instances started in the fleet.

NoProtection – The game session can be terminated during a scale-down event.

FullProtection – If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

operating_system: Option<String>

Operating system of the fleet's computing resources. A fleet's operating system depends on the OS specified for the build that is deployed on this fleet.

resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>

Fleet policy to limit the number of game sessions an individual player can create over a span of time.

server_launch_parameters: Option<String>

Game server launch parameters specified for fleets created prior to 2016-08-04 (or AWS SDK v. 0.12.16). Server launch parameters for fleets created after this date are specified in the fleet's RuntimeConfiguration.

server_launch_path: Option<String>

Path to a game server executable in the fleet's build, specified for fleets created prior to 2016-08-04 (or AWS SDK v. 0.12.16). Server launch paths for fleets created after this date are specified in the fleet's RuntimeConfiguration.

status: Option<String>

Current status of the fleet.

Possible fleet statuses include the following:

NEW – A new fleet has been defined and desired instances is set to 1.

DOWNLOADING/VALIDATING/BUILDING/ACTIVATING – Amazon GameLift is setting up the new fleet, creating new instances with the game build and starting server processes.

ACTIVE – Hosts can now accept game sessions.

ERROR – An error occurred when downloading, validating, building, or activating the fleet.

DELETING – Hosts are responding to a delete fleet request.

TERMINATED – The fleet no longer exists.

termination_time: Option<f64>

Time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").