Struct rusoto_gamelift::GameLiftClient

pub struct GameLiftClient { /* fields omitted */ }

A client for the Amazon GameLift API.

Methods

impl GameLiftClient

pub fn new(region: Region) -> GameLiftClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> GameLiftClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,