Struct rusoto_gamelift::GameSessionPlacement

pub struct GameSessionPlacement {
    pub end_time: Option<f64>,
    pub game_properties: Option<Vec<GameProperty>>,
    pub game_session_arn: Option<String>,
    pub game_session_id: Option<String>,
    pub game_session_name: Option<String>,
    pub game_session_queue_name: Option<String>,
    pub game_session_region: Option<String>,
    pub ip_address: Option<String>,
    pub maximum_player_session_count: Option<i64>,
    pub placed_player_sessions: Option<Vec<PlacedPlayerSession>>,
    pub placement_id: Option<String>,
    pub player_latencies: Option<Vec<PlayerLatency>>,
    pub port: Option<i64>,
    pub start_time: Option<f64>,
    pub status: Option<String>,
}

Object that describes a StartGameSessionPlacement request. This object includes the full details of the original request plus the current status and start/end time stamps.

Game session placement-related operations include:

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

Fields
end_time: Option<f64>

Time stamp indicating when this request was completed, canceled, or timed out.

game_properties: Option<Vec<GameProperty>>

Set of developer-defined properties for a game session. These properties are passed to the server process hosting the game session.

game_session_arn: Option<String>

Identifier for the game session created by this placement request. This value is set once the new game session is placed (placement status is Fulfilled). This identifier is unique across all regions. You can use this value as a GameSessionId value as needed.

game_session_id: Option<String>

Unique identifier for the game session. This value is set once the new game session is placed (placement status is Fulfilled).

game_session_name: Option<String>

Descriptive label that is associated with a game session. Session names do not need to be unique.

game_session_queue_name: Option<String>

Descriptive label that is associated with queue. Queue names must be unique within each region.

game_session_region: Option<String>

Name of the region where the game session created by this placement request is running. This value is set once the new game session is placed (placement status is Fulfilled).

ip_address: Option<String>

IP address of the game session. To connect to a Amazon GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is Fulfilled).

maximum_player_session_count: Option<i64>

Maximum number of players that can be connected simultaneously to the game session.

placed_player_sessions: Option<Vec<PlacedPlayerSession>>

Collection of information on player sessions created in response to the game session placement request. These player sessions are created only once a new game session is successfully placed (placement status is Fulfilled). This information includes the player ID (as provided in the placement request) and the corresponding player session ID. Retrieve full player sessions by calling DescribePlayerSessions with the player session ID.

placement_id: Option<String>

Unique identifier for a game session placement.

player_latencies: Option<Vec<PlayerLatency>>

Set of values, expressed in milliseconds, indicating the amount of latency that players are experiencing when connected to AWS regions.

port: Option<i64>

Port number for the game session. To connect to a Amazon GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is Fulfilled).

start_time: Option<f64>

Time stamp indicating when this request was placed in the queue. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status: Option<String>

Current status of the game session placement request.

PENDING – The placement request is currently in the queue waiting to be processed.

FULFILLED – A new game session and player sessions (if requested) have been successfully created. Values for GameSessionArn and GameSessionRegion are available.

CANCELLED – The placement request was canceled with a call to StopGameSessionPlacement.

TIMED_OUT – A new game session was not successfully created before the time limit expired. You can resubmit the placement request as needed.