Struct rusoto_gamelift::Instance

pub struct Instance {
    pub creation_time: Option<f64>,
    pub fleet_id: Option<String>,
    pub instance_id: Option<String>,
    pub ip_address: Option<String>,
    pub operating_system: Option<String>,
    pub status: Option<String>,
    pub type_: Option<String>,
}

Properties that describe an instance of a virtual computing resource that hosts one or more game servers. A fleet contains zero or more instances.

Fields
creation_time: Option<f64>

Time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

fleet_id: Option<String>

Unique identifier for a fleet that the instance is in.

instance_id: Option<String>

Unique identifier for an instance.

ip_address: Option<String>

IP address assigned to the instance.

operating_system: Option<String>

Operating system that is running on this instance.

status: Option<String>

Current status of the instance. Possible statuses include the following:

PENDING – The instance is in the process of being created and launching server processes as defined in the fleet's runtime configuration.

ACTIVE – The instance has been successfully created and at least one server process has successfully launched and reported back to Amazon GameLift that it is ready to host a game session. The instance is now considered ready to host game sessions.

TERMINATING – The instance is in the process of shutting down. This may happen to reduce capacity during a scaling down event or to recycle resources in the event of a problem.

type_: Option<String>

EC2 instance type that defines the computing resources of this instance.