Struct rusoto_gamelift::MatchmakingConfiguration

pub struct MatchmakingConfiguration {
    pub acceptance_required: Option<bool>,
    pub acceptance_timeout_seconds: Option<i64>,
    pub additional_player_count: Option<i64>,
    pub creation_time: Option<f64>,
    pub custom_event_data: Option<String>,
    pub description: Option<String>,
    pub game_properties: Option<Vec<GameProperty>>,
    pub game_session_data: Option<String>,
    pub game_session_queue_arns: Option<Vec<String>>,
    pub name: Option<String>,
    pub notification_target: Option<String>,
    pub request_timeout_seconds: Option<i64>,
    pub rule_set_name: Option<String>,
}

Guidelines for use with FlexMatch to match players into games. All matchmaking requests must specify a matchmaking configuration.

Fields
acceptance_required: Option<bool>

Flag that determines whether or not a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE.

acceptance_timeout_seconds: Option<i64>

Length of time (in seconds) to wait for players to accept a proposed match. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

additional_player_count: Option<i64>

Number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match.

creation_time: Option<f64>

Time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

custom_event_data: Option<String>

Information to attached to all events related to the matchmaking configuration.

description: Option<String>

Descriptive label that is associated with matchmaking configuration.

game_properties: Option<Vec<GameProperty>>

Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match.

game_session_data: Option<String>

Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match.

game_session_queue_arns: Option<Vec<String>>

Amazon Resource Name (ARN) that is assigned to a game session queue and uniquely identifies it. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. These queues are used when placing game sessions for matches that are created with this matchmaking configuration. Queues can be located in any region.

name: Option<String>

Unique identifier for a matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

notification_target: Option<String>

SNS topic ARN that is set up to receive matchmaking notifications.

request_timeout_seconds: Option<i64>

Maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that time out can be resubmitted as needed.

rule_set_name: Option<String>

Unique identifier for a matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same region.