Struct rusoto_gamelift::MatchmakingTicket

pub struct MatchmakingTicket {
    pub configuration_name: Option<String>,
    pub game_session_connection_info: Option<GameSessionConnectionInfo>,
    pub players: Option<Vec<Player>>,
    pub start_time: Option<f64>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub status_reason: Option<String>,
    pub ticket_id: Option<String>,
}

Ticket generated to track the progress of a matchmaking request. Each ticket is uniquely identified by a ticket ID, supplied by the requester, when creating a matchmaking request with StartMatchmaking. Tickets can be retrieved by calling DescribeMatchmaking with the ticket ID.

Fields
configuration_name: Option<String>

Name of the MatchmakingConfiguration that is used with this ticket. Matchmaking configurations determine how players are grouped into a match and how a new game session is created for the match.

game_session_connection_info: Option<GameSessionConnectionInfo>

Identifier and connection information of the game session created for the match. This information is added to the ticket only after the matchmaking request has been successfully completed.

players: Option<Vec<Player>>

A set of Player objects, each representing a player to find matches for. Players are identified by a unique player ID and may include latency data for use during matchmaking. If the ticket is in status COMPLETED, the Player objects include the team the players were assigned to in the resulting match.

start_time: Option<f64>

Time stamp indicating when this matchmaking request was received. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status: Option<String>

Current status of the matchmaking request.

QUEUED – The matchmaking request has been received and is currently waiting to be processed.

SEARCHING – The matchmaking request is currently being processed.

REQUIRES_ACCEPTANCE – A match has been proposed and the players must accept the match (see AcceptMatch). This status is used only with requests that use a matchmaking configuration with a player acceptance requirement.

PLACING – The FlexMatch engine has matched players and is in the process of placing a new game session for the match.

COMPLETED – Players have been matched and a game session is ready to host the players. A ticket in this state contains the necessary connection information for players.

FAILED – The matchmaking request was not completed. Tickets with players who fail to accept a proposed match are placed in FAILED status; new matchmaking requests can be submitted for these players.

CANCELLED – The matchmaking request was canceled with a call to StopMatchmaking.

TIMED_OUT – The matchmaking request was not completed within the duration specified in the matchmaking configuration. Matchmaking requests that time out can be resubmitted.

status_message: Option<String>

Additional information about the current status.

status_reason: Option<String>

Code to explain the current status. For example, a status reason may indicate when a ticket has returned to SEARCHING status after a proposed match fails to receive player acceptances.

ticket_id: Option<String>

Unique identifier for a matchmaking ticket.