Struct rusoto_gamelift::RuntimeConfiguration

pub struct RuntimeConfiguration {
    pub game_session_activation_timeout_seconds: Option<i64>,
    pub max_concurrent_game_session_activations: Option<i64>,
    pub server_processes: Option<Vec<ServerProcess>>,
}

Collection of server process configurations that describe what processes should be run on each instance in a fleet. An instance can launch and maintain multiple server processes based on the runtime configuration; it regularly checks for an updated runtime configuration and starts new server processes to match the latest version.

The key purpose of a runtime configuration with multiple server process configurations is to be able to run more than one kind of game server in a single fleet. You can include configurations for more than one server executable in order to run two or more different programs to run on the same instance. This option might be useful, for example, to run more than one version of your game server on the same fleet. Another option is to specify configurations for the same server executable but with different launch parameters.

A Amazon GameLift instance is limited to 50 processes running simultaneously. To calculate the total number of processes specified in a runtime configuration, add the values of the ConcurrentExecutions parameter for each ServerProcess object in the runtime configuration.

Fields
game_session_activation_timeout_seconds: Option<i64>

Maximum amount of time (in seconds) that a game session can remain in status ACTIVATING. If the game session is not active before the timeout, activation is terminated and the game session status is changed to TERMINATED.

max_concurrent_game_session_activations: Option<i64>

Maximum number of game sessions with status ACTIVATING to allow on an instance simultaneously. This setting limits the amount of instance resources that can be used for new game activations at any one time.

server_processes: Option<Vec<ServerProcess>>

Collection of server process configurations that describe which server processes to run on each instance in a fleet.