Struct rusoto_gamelift::ScalingPolicy

pub struct ScalingPolicy {
    pub comparison_operator: Option<String>,
    pub evaluation_periods: Option<i64>,
    pub fleet_id: Option<String>,
    pub metric_name: Option<String>,
    pub name: Option<String>,
    pub scaling_adjustment: Option<i64>,
    pub scaling_adjustment_type: Option<String>,
    pub status: Option<String>,
    pub threshold: Option<f64>,
}

Rule that controls how a fleet is scaled. Scaling policies are uniquely identified by the combination of name and fleet ID.

Fields
comparison_operator: Option<String>

Comparison operator to use when measuring a metric against the threshold value.

evaluation_periods: Option<i64>

Length of time (in minutes) the metric must be at or beyond the threshold before a scaling event is triggered.

fleet_id: Option<String>

Unique identifier for a fleet that is associated with this scaling policy.

metric_name: Option<String>

Name of the Amazon GameLift-defined metric that is used to trigger an adjustment.

ActivatingGameSessions – number of game sessions in the process of being created (game session status = ACTIVATING).

ActiveGameSessions – number of game sessions currently running (game session status = ACTIVE).

CurrentPlayerSessions – number of active or reserved player sessions (player session status = ACTIVE or RESERVED).

AvailablePlayerSessions – number of player session slots currently available in active game sessions across the fleet, calculated by subtracting a game session's current player session count from its maximum player session count. This number does include game sessions that are not currently accepting players (game session PlayerSessionCreationPolicy = DENY_ALL).

ActiveInstances – number of instances currently running a game session.

IdleInstances – number of instances not currently running a game session.

name: Option<String>

Descriptive label that is associated with a scaling policy. Policy names do not need to be unique.

scaling_adjustment: Option<i64>

Amount of adjustment to make, based on the scaling adjustment type.

scaling_adjustment_type: Option<String>

Type of adjustment to make to a fleet's instance count (see FleetCapacity):

ChangeInCapacity – add (or subtract) the scaling adjustment value from the current instance count. Positive values scale up while negative values scale down.

ExactCapacity – set the instance count to the scaling adjustment value.

PercentChangeInCapacity – increase or reduce the current instance count by the scaling adjustment, read as a percentage. Positive values scale up while negative values scale down.

status: Option<String>

Current status of the scaling policy. The scaling policy is only in force when in an ACTIVE status.

ACTIVE – The scaling policy is currently in force.

UPDATE_REQUESTED – A request to update the scaling policy has been received.

UPDATING – A change is being made to the scaling policy.

DELETE_REQUESTED – A request to delete the scaling policy has been received.

DELETING – The scaling policy is being deleted.

DELETED – The scaling policy has been deleted.

ERROR – An error occurred in creating the policy. It should be removed and recreated.

threshold: Option<f64>

Metric value used to trigger a scaling event.