Trait rusoto_gamelift::GameLift

pub trait GameLift {
    fn create_alias(
        &self, 
        input: &CreateAliasInput
    ) -> Result<CreateAliasOutput, CreateAliasError>;
    fn create_build(
        &self, 
        input: &CreateBuildInput
    ) -> Result<CreateBuildOutput, CreateBuildError>;
    fn create_fleet(
        &self, 
        input: &CreateFleetInput
    ) -> Result<CreateFleetOutput, CreateFleetError>;
    fn create_game_session(
        &self, 
        input: &CreateGameSessionInput
    ) -> Result<CreateGameSessionOutput, CreateGameSessionError>;
    fn create_game_session_queue(
        &self, 
        input: &CreateGameSessionQueueInput
    ) -> Result<CreateGameSessionQueueOutput, CreateGameSessionQueueError>;
    fn create_player_session(
        &self, 
        input: &CreatePlayerSessionInput
    ) -> Result<CreatePlayerSessionOutput, CreatePlayerSessionError>;
    fn create_player_sessions(
        &self, 
        input: &CreatePlayerSessionsInput
    ) -> Result<CreatePlayerSessionsOutput, CreatePlayerSessionsError>;
    fn delete_alias(
        &self, 
        input: &DeleteAliasInput
    ) -> Result<(), DeleteAliasError>;
    fn delete_build(
        &self, 
        input: &DeleteBuildInput
    ) -> Result<(), DeleteBuildError>;
    fn delete_fleet(
        &self, 
        input: &DeleteFleetInput
    ) -> Result<(), DeleteFleetError>;
    fn delete_game_session_queue(
        &self, 
        input: &DeleteGameSessionQueueInput
    ) -> Result<DeleteGameSessionQueueOutput, DeleteGameSessionQueueError>;
    fn delete_scaling_policy(
        &self, 
        input: &DeleteScalingPolicyInput
    ) -> Result<(), DeleteScalingPolicyError>;
    fn describe_alias(
        &self, 
        input: &DescribeAliasInput
    ) -> Result<DescribeAliasOutput, DescribeAliasError>;
    fn describe_build(
        &self, 
        input: &DescribeBuildInput
    ) -> Result<DescribeBuildOutput, DescribeBuildError>;
    fn describe_ec2_instance_limits(
        &self, 
        input: &DescribeEC2InstanceLimitsInput
    ) -> Result<DescribeEC2InstanceLimitsOutput, DescribeEC2InstanceLimitsError>;
    fn describe_fleet_attributes(
        &self, 
        input: &DescribeFleetAttributesInput
    ) -> Result<DescribeFleetAttributesOutput, DescribeFleetAttributesError>;
    fn describe_fleet_capacity(
        &self, 
        input: &DescribeFleetCapacityInput
    ) -> Result<DescribeFleetCapacityOutput, DescribeFleetCapacityError>;
    fn describe_fleet_events(
        &self, 
        input: &DescribeFleetEventsInput
    ) -> Result<DescribeFleetEventsOutput, DescribeFleetEventsError>;
    fn describe_fleet_port_settings(
        &self, 
        input: &DescribeFleetPortSettingsInput
    ) -> Result<DescribeFleetPortSettingsOutput, DescribeFleetPortSettingsError>;
    fn describe_fleet_utilization(
        &self, 
        input: &DescribeFleetUtilizationInput
    ) -> Result<DescribeFleetUtilizationOutput, DescribeFleetUtilizationError>;
    fn describe_game_session_details(
        &self, 
        input: &DescribeGameSessionDetailsInput
    ) -> Result<DescribeGameSessionDetailsOutput, DescribeGameSessionDetailsError>;
    fn describe_game_session_placement(
        &self, 
        input: &DescribeGameSessionPlacementInput
    ) -> Result<DescribeGameSessionPlacementOutput, DescribeGameSessionPlacementError>;
    fn describe_game_session_queues(
        &self, 
        input: &DescribeGameSessionQueuesInput
    ) -> Result<DescribeGameSessionQueuesOutput, DescribeGameSessionQueuesError>;
    fn describe_game_sessions(
        &self, 
        input: &DescribeGameSessionsInput
    ) -> Result<DescribeGameSessionsOutput, DescribeGameSessionsError>;
    fn describe_instances(
        &self, 
        input: &DescribeInstancesInput
    ) -> Result<DescribeInstancesOutput, DescribeInstancesError>;
    fn describe_player_sessions(
        &self, 
        input: &DescribePlayerSessionsInput
    ) -> Result<DescribePlayerSessionsOutput, DescribePlayerSessionsError>;
    fn describe_runtime_configuration(
        &self, 
        input: &DescribeRuntimeConfigurationInput
    ) -> Result<DescribeRuntimeConfigurationOutput, DescribeRuntimeConfigurationError>;
    fn describe_scaling_policies(
        &self, 
        input: &DescribeScalingPoliciesInput
    ) -> Result<DescribeScalingPoliciesOutput, DescribeScalingPoliciesError>;
    fn get_game_session_log_url(
        &self, 
        input: &GetGameSessionLogUrlInput
    ) -> Result<GetGameSessionLogUrlOutput, GetGameSessionLogUrlError>;
    fn get_instance_access(
        &self, 
        input: &GetInstanceAccessInput
    ) -> Result<GetInstanceAccessOutput, GetInstanceAccessError>;
    fn list_aliases(
        &self, 
        input: &ListAliasesInput
    ) -> Result<ListAliasesOutput, ListAliasesError>;
    fn list_builds(
        &self, 
        input: &ListBuildsInput
    ) -> Result<ListBuildsOutput, ListBuildsError>;
    fn list_fleets(
        &self, 
        input: &ListFleetsInput
    ) -> Result<ListFleetsOutput, ListFleetsError>;
    fn put_scaling_policy(
        &self, 
        input: &PutScalingPolicyInput
    ) -> Result<PutScalingPolicyOutput, PutScalingPolicyError>;
    fn request_upload_credentials(
        &self, 
        input: &RequestUploadCredentialsInput
    ) -> Result<RequestUploadCredentialsOutput, RequestUploadCredentialsError>;
    fn resolve_alias(
        &self, 
        input: &ResolveAliasInput
    ) -> Result<ResolveAliasOutput, ResolveAliasError>;
    fn search_game_sessions(
        &self, 
        input: &SearchGameSessionsInput
    ) -> Result<SearchGameSessionsOutput, SearchGameSessionsError>;
    fn start_game_session_placement(
        &self, 
        input: &StartGameSessionPlacementInput
    ) -> Result<StartGameSessionPlacementOutput, StartGameSessionPlacementError>;
    fn stop_game_session_placement(
        &self, 
        input: &StopGameSessionPlacementInput
    ) -> Result<StopGameSessionPlacementOutput, StopGameSessionPlacementError>;
    fn update_alias(
        &self, 
        input: &UpdateAliasInput
    ) -> Result<UpdateAliasOutput, UpdateAliasError>;
    fn update_build(
        &self, 
        input: &UpdateBuildInput
    ) -> Result<UpdateBuildOutput, UpdateBuildError>;
    fn update_fleet_attributes(
        &self, 
        input: &UpdateFleetAttributesInput
    ) -> Result<UpdateFleetAttributesOutput, UpdateFleetAttributesError>;
    fn update_fleet_capacity(
        &self, 
        input: &UpdateFleetCapacityInput
    ) -> Result<UpdateFleetCapacityOutput, UpdateFleetCapacityError>;
    fn update_fleet_port_settings(
        &self, 
        input: &UpdateFleetPortSettingsInput
    ) -> Result<UpdateFleetPortSettingsOutput, UpdateFleetPortSettingsError>;
    fn update_game_session(
        &self, 
        input: &UpdateGameSessionInput
    ) -> Result<UpdateGameSessionOutput, UpdateGameSessionError>;
    fn update_game_session_queue(
        &self, 
        input: &UpdateGameSessionQueueInput
    ) -> Result<UpdateGameSessionQueueOutput, UpdateGameSessionQueueError>;
    fn update_runtime_configuration(
        &self, 
        input: &UpdateRuntimeConfigurationInput
    ) -> Result<UpdateRuntimeConfigurationOutput, UpdateRuntimeConfigurationError>;
}

Trait representing the capabilities of the Amazon GameLift API. Amazon GameLift clients implement this trait.

Required Methods

fn create_alias(
&self,
input: &CreateAliasInput
) -> Result<CreateAliasOutput, CreateAliasError>

Creates an alias and sets a target fleet. A fleet alias can be used in place of a fleet ID, such as when calling CreateGameSession from a game client or game service or adding destinations to a game session queue. By changing an alias's target fleet, you can switch your players to the new fleet without changing any other component. In production, this feature is particularly useful to redirect your player base seamlessly to the latest game server update.

Amazon GameLift supports two types of routing strategies for aliases: simple and terminal. Use a simple alias to point to an active fleet. Use a terminal alias to display a message to incoming traffic instead of routing players to an active fleet. This option is useful when a game server is no longer supported but you want to provide better messaging than a standard 404 error.

To create a fleet alias, specify an alias name, routing strategy, and optional description. If successful, a new alias record is returned, including an alias ID, which you can reference when creating a game session. To reassign the alias to another fleet ID, call UpdateAlias.

fn create_build(
&self,
input: &CreateBuildInput
) -> Result<CreateBuildOutput, CreateBuildError>

Creates a new Amazon GameLift build from a set of game server binary files stored in an Amazon Simple Storage Service (Amazon S3) location. When using this API call, you must create a .zip file containing all of the build files and store it in an Amazon S3 bucket under your AWS account. For help on packaging your build files and creating a build, see Uploading Your Game to Amazon GameLift.

Use this API action ONLY if you are storing your game build files in an Amazon S3 bucket in your AWS account. To create a build using files stored in a directory, use the CLI command upload-build , which uploads the build files from a file location you specify and creates a build.

To create a new build using CreateBuild, identify the storage location and operating system of your game build. You also have the option of specifying a build name and version. If successful, this action creates a new build record with an unique build ID and in INITIALIZED status. Use the API call DescribeBuild to check the status of your build. A build must be in READY status before it can be used to create fleets to host your game.

fn create_fleet(
&self,
input: &CreateFleetInput
) -> Result<CreateFleetOutput, CreateFleetError>

Creates a new fleet to run your game servers. A fleet is a set of Amazon Elastic Compute Cloud (Amazon EC2) instances, each of which can run multiple server processes to host game sessions. You configure a fleet to create instances with certain hardware specifications (see Amazon EC2 Instance Types for more information), and deploy a specified game build to each instance. A newly created fleet passes through several statuses; once it reaches the ACTIVE status, it can begin hosting game sessions.

To create a new fleet, you must specify the following: (1) fleet name, (2) build ID of an uploaded game build, (3) an EC2 instance type, and (4) a runtime configuration that describes which server processes to run on each instance in the fleet. (Although the runtime configuration is not a required parameter, the fleet cannot be successfully created without it.) You can also configure the new fleet with the following settings: fleet description, access permissions for inbound traffic, fleet-wide game session protection, and resource creation limit. If you use Amazon CloudWatch for metrics, you can add the new fleet to a metric group, which allows you to view aggregated metrics for a set of fleets. Once you specify a metric group, the new fleet's metrics are included in the metric group's data.

If the CreateFleet call is successful, Amazon GameLift performs the following tasks:

Creates a fleet record and sets the status to NEW (followed by other statuses as the fleet is activated).

Sets the fleet's capacity to 1 "desired", which causes Amazon GameLift to start one new EC2 instance.

Starts launching server processes on the instance. If the fleet is configured to run multiple server processes per instance, Amazon GameLift staggers each launch by a few seconds.

Begins writing events to the fleet event log, which can be accessed in the Amazon GameLift console.

Sets the fleet's status to ACTIVE once one server process in the fleet is ready to host a game session.

After a fleet is created, use the following actions to change fleet properties and configuration:

UpdateFleetAttributes -- Update fleet metadata, including name and description.

UpdateFleetCapacity -- Increase or decrease the number of instances you want the fleet to maintain.

UpdateFleetPortSettings -- Change the IP address and port ranges that allow access to incoming traffic.

UpdateRuntimeConfiguration -- Change how server processes are launched in the fleet, including launch path, launch parameters, and the number of concurrent processes.

PutScalingPolicy -- Create or update rules that are used to set the fleet's capacity (autoscaling).

fn create_game_session(
&self,
input: &CreateGameSessionInput
) -> Result<CreateGameSessionOutput, CreateGameSessionError>

Creates a multiplayer game session for players. This action creates a game session record and assigns an available server process in the specified fleet to host the game session. A fleet must have an ACTIVE status before a game session can be created in it.

To create a game session, specify either fleet ID or alias ID and indicate a maximum number of players to allow in the game session. You can also provide a name and game-specific properties for this game session. If successful, a GameSession object is returned containing game session properties, including a game session ID with the custom string you provided.

Idempotency tokens. You can add a token that uniquely identifies game session requests. This is useful for ensuring that game session requests are idempotent. Multiple requests with the same idempotency token are processed only once; subsequent requests return the original result. All response values are the same with the exception of game session status, which may change.

Resource creation limits. If you are creating a game session on a fleet with a resource creation limit policy in force, then you must specify a creator ID. Without this ID, Amazon GameLift has no way to evaluate the policy for this new game session request.

By default, newly created game sessions allow new players to join. Use UpdateGameSession to change the game session's player session creation policy.

Available in Amazon GameLift Local.

fn create_game_session_queue(
&self,
input: &CreateGameSessionQueueInput
) -> Result<CreateGameSessionQueueOutput, CreateGameSessionQueueError>

Establishes a new queue for processing requests to place new game sessions. A queue identifies where new game sessions can be hosted -- by specifying a list of destinations (fleets or aliases) -- and how long requests can wait in the queue before timing out. You can set up a queue to try to place game sessions on fleets in multiple regions. To add placement requests to a queue, call StartGameSessionPlacement and reference the queue name.

Destination order. When processing a request for a game session, Amazon GameLift tries each destination in order until it finds one with available resources to host the new game session. A queue's default order is determined by how destinations are listed. The default order is overridden when a game session placement request provides player latency information. Player latency information enables Amazon GameLift to prioritize destinations where players report the lowest average latency, as a result placing the new game session where the majority of players will have the best possible gameplay experience.

Player latency policies. For placement requests containing player latency information, use player latency policies to protect individual players from very high latencies. With a latency cap, even when a destination can deliver a low latency for most players, the game is not placed where any individual player is reporting latency higher than a policy's maximum. A queue can have multiple latency policies, which are enforced consecutively starting with the policy with the lowest latency cap. Use multiple policies to gradually relax latency controls; for example, you might set a policy with a low latency cap for the first 60 seconds, a second policy with a higher cap for the next 60 seconds, etc.

To create a new queue, provide a name, timeout value, a list of destinations and, if desired, a set of latency policies. If successful, a new queue object is returned.

fn create_player_session(
&self,
input: &CreatePlayerSessionInput
) -> Result<CreatePlayerSessionOutput, CreatePlayerSessionError>

Adds a player to a game session and creates a player session record. Before a player can be added, a game session must have an ACTIVE status, have a creation policy of ALLOW_ALL, and have an open player slot. To add a group of players to a game session, use CreatePlayerSessions.

To create a player session, specify a game session ID, player ID, and optionally a string of player data. If successful, the player is added to the game session and a new PlayerSession object is returned. Player sessions cannot be updated.

Available in Amazon GameLift Local.

fn create_player_sessions(
&self,
input: &CreatePlayerSessionsInput
) -> Result<CreatePlayerSessionsOutput, CreatePlayerSessionsError>

Adds a group of players to a game session. This action is useful with a team matching feature. Before players can be added, a game session must have an ACTIVE status, have a creation policy of ALLOW_ALL, and have an open player slot. To add a single player to a game session, use CreatePlayerSession.

To create player sessions, specify a game session ID, a list of player IDs, and optionally a set of player data strings. If successful, the players are added to the game session and a set of new PlayerSession objects is returned. Player sessions cannot be updated.

Available in Amazon GameLift Local.

fn delete_alias(&self, input: &DeleteAliasInput) -> Result<(), DeleteAliasError>

Deletes a fleet alias. This action removes all record of the alias. Game clients attempting to access a server process using the deleted alias receive an error. To delete an alias, specify the alias ID to be deleted.

fn delete_build(&self, input: &DeleteBuildInput) -> Result<(), DeleteBuildError>

Deletes a build. This action permanently deletes the build record and any uploaded build files.

To delete a build, specify its ID. Deleting a build does not affect the status of any active fleets using the build, but you can no longer create new fleets with the deleted build.

fn delete_fleet(&self, input: &DeleteFleetInput) -> Result<(), DeleteFleetError>

Deletes everything related to a fleet. Before deleting a fleet, you must set the fleet's desired capacity to zero. See UpdateFleetCapacity.

This action removes the fleet's resources and the fleet record. Once a fleet is deleted, you can no longer use that fleet.

fn delete_game_session_queue(
&self,
input: &DeleteGameSessionQueueInput
) -> Result<DeleteGameSessionQueueOutput, DeleteGameSessionQueueError>

Deletes a game session queue. This action means that any StartGameSessionPlacement requests that reference this queue will fail. To delete a queue, specify the queue name.

fn delete_scaling_policy(
&self,
input: &DeleteScalingPolicyInput
) -> Result<(), DeleteScalingPolicyError>

Deletes a fleet scaling policy. This action means that the policy is no longer in force and removes all record of it. To delete a scaling policy, specify both the scaling policy name and the fleet ID it is associated with.

fn describe_alias(
&self,
input: &DescribeAliasInput
) -> Result<DescribeAliasOutput, DescribeAliasError>

Retrieves properties for a fleet alias. This operation returns all alias metadata and settings. To get just the fleet ID an alias is currently pointing to, use ResolveAlias.

To get alias properties, specify the alias ID. If successful, an Alias object is returned.

fn describe_build(
&self,
input: &DescribeBuildInput
) -> Result<DescribeBuildOutput, DescribeBuildError>

Retrieves properties for a build. To get a build record, specify a build ID. If successful, an object containing the build properties is returned.

fn describe_ec2_instance_limits(
&self,
input: &DescribeEC2InstanceLimitsInput
) -> Result<DescribeEC2InstanceLimitsOutput, DescribeEC2InstanceLimitsError>

Retrieves the following information for the specified EC2 instance type:

maximum number of instances allowed per AWS account (service limit)

current usage level for the AWS account

Service limits vary depending on region. Available regions for Amazon GameLift can be found in the AWS Management Console for Amazon GameLift (see the drop-down list in the upper right corner).

fn describe_fleet_attributes(
&self,
input: &DescribeFleetAttributesInput
) -> Result<DescribeFleetAttributesOutput, DescribeFleetAttributesError>

Retrieves fleet properties, including metadata, status, and configuration, for one or more fleets. You can request attributes for all fleets, or specify a list of one or more fleet IDs. When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a FleetAttributes object is returned for each requested fleet ID. When specifying a list of fleet IDs, attribute objects are returned only for fleets that currently exist.

Some API actions may limit the number of fleet IDs allowed in one request. If a request exceeds this limit, the request fails and the error message includes the maximum allowed.

fn describe_fleet_capacity(
&self,
input: &DescribeFleetCapacityInput
) -> Result<DescribeFleetCapacityOutput, DescribeFleetCapacityError>

Retrieves the current status of fleet capacity for one or more fleets. This information includes the number of instances that have been requested for the fleet and the number currently active. You can request capacity for all fleets, or specify a list of one or more fleet IDs. When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a FleetCapacity object is returned for each requested fleet ID. When specifying a list of fleet IDs, attribute objects are returned only for fleets that currently exist.

Some API actions may limit the number of fleet IDs allowed in one request. If a request exceeds this limit, the request fails and the error message includes the maximum allowed.

fn describe_fleet_events(
&self,
input: &DescribeFleetEventsInput
) -> Result<DescribeFleetEventsOutput, DescribeFleetEventsError>

Retrieves entries from the specified fleet's event log. You can specify a time range to limit the result set. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a collection of event log entries matching the request are returned.

fn describe_fleet_port_settings(
&self,
input: &DescribeFleetPortSettingsInput
) -> Result<DescribeFleetPortSettingsOutput, DescribeFleetPortSettingsError>

Retrieves the inbound connection permissions for a fleet. Connection permissions include a range of IP addresses and port settings that incoming traffic can use to access server processes in the fleet. To get a fleet's inbound connection permissions, specify a fleet ID. If successful, a collection of IpPermission objects is returned for the requested fleet ID. If the requested fleet has been deleted, the result set is empty.

fn describe_fleet_utilization(
&self,
input: &DescribeFleetUtilizationInput
) -> Result<DescribeFleetUtilizationOutput, DescribeFleetUtilizationError>

Retrieves utilization statistics for one or more fleets. You can request utilization data for all fleets, or specify a list of one or more fleet IDs. When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a FleetUtilization object is returned for each requested fleet ID. When specifying a list of fleet IDs, utilization objects are returned only for fleets that currently exist.

Some API actions may limit the number of fleet IDs allowed in one request. If a request exceeds this limit, the request fails and the error message includes the maximum allowed.

fn describe_game_session_details(
&self,
input: &DescribeGameSessionDetailsInput
) -> Result<DescribeGameSessionDetailsOutput, DescribeGameSessionDetailsError>

Retrieves properties, including the protection policy in force, for one or more game sessions. This action can be used in several ways: (1) provide a GameSessionId or GameSessionArn to request details for a specific game session; (2) provide either a FleetId or an AliasId to request properties for all game sessions running on a fleet.

To get game session record(s), specify just one of the following: game session ID, fleet ID, or alias ID. You can filter this request by game session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSessionDetail object is returned for each session matching the request.

fn describe_game_session_placement(
&self,
input: &DescribeGameSessionPlacementInput
) -> Result<DescribeGameSessionPlacementOutput, DescribeGameSessionPlacementError>

Retrieves properties and current status of a game session placement request. To get game session placement details, specify the placement ID. If successful, a GameSessionPlacement object is returned.

fn describe_game_session_queues(
&self,
input: &DescribeGameSessionQueuesInput
) -> Result<DescribeGameSessionQueuesOutput, DescribeGameSessionQueuesError>

Retrieves the properties for one or more game session queues. When requesting multiple queues, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSessionQueue object is returned for each requested queue. When specifying a list of queues, objects are returned only for queues that currently exist in the region.

fn describe_game_sessions(
&self,
input: &DescribeGameSessionsInput
) -> Result<DescribeGameSessionsOutput, DescribeGameSessionsError>

Retrieves a set of one or more game sessions. Request a specific game session or request all game sessions on a fleet. Alternatively, use SearchGameSessions to request a set of active game sessions that are filtered by certain criteria. To retrieve protection policy settings for game sessions, use DescribeGameSessionDetails.

To get game sessions, specify one of the following: game session ID, fleet ID, or alias ID. You can filter this request by game session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSession object is returned for each game session matching the request.

Available in Amazon GameLift Local.

fn describe_instances(
&self,
input: &DescribeInstancesInput
) -> Result<DescribeInstancesOutput, DescribeInstancesError>

Retrieves information about a fleet's instances, including instance IDs. Use this action to get details on all instances in the fleet or get details on one specific instance.

To get a specific instance, specify fleet ID and instance ID. To get all instances in a fleet, specify a fleet ID only. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, an Instance object is returned for each result.

fn describe_player_sessions(
&self,
input: &DescribePlayerSessionsInput
) -> Result<DescribePlayerSessionsOutput, DescribePlayerSessionsError>

Retrieves properties for one or more player sessions. This action can be used in several ways: (1) provide a PlayerSessionId to request properties for a specific player session; (2) provide a GameSessionId to request properties for all player sessions in the specified game session; (3) provide a PlayerId to request properties for all player sessions of a specified player.

To get game session record(s), specify only one of the following: a player session ID, a game session ID, or a player ID. You can filter this request by player session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a PlayerSession object is returned for each session matching the request.

Available in Amazon GameLift Local.

fn describe_runtime_configuration(
&self,
input: &DescribeRuntimeConfigurationInput
) -> Result<DescribeRuntimeConfigurationOutput, DescribeRuntimeConfigurationError>

Retrieves the current runtime configuration for the specified fleet. The runtime configuration tells Amazon GameLift how to launch server processes on instances in the fleet.

fn describe_scaling_policies(
&self,
input: &DescribeScalingPoliciesInput
) -> Result<DescribeScalingPoliciesOutput, DescribeScalingPoliciesError>

Retrieves all scaling policies applied to a fleet.

To get a fleet's scaling policies, specify the fleet ID. You can filter this request by policy status, such as to retrieve only active scaling policies. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, set of ScalingPolicy objects is returned for the fleet.

fn get_game_session_log_url(
&self,
input: &GetGameSessionLogUrlInput
) -> Result<GetGameSessionLogUrlOutput, GetGameSessionLogUrlError>

Retrieves the location of stored game session logs for a specified game session. When a game session is terminated, Amazon GameLift automatically stores the logs in Amazon S3. Use this URL to download the logs.

See the AWS Service Limits page for maximum log file sizes. Log files that exceed this limit are not saved.

fn get_instance_access(
&self,
input: &GetInstanceAccessInput
) -> Result<GetInstanceAccessOutput, GetInstanceAccessError>

Requests remote access to a fleet instance. Remote access is useful for debugging, gathering benchmarking data, or watching activity in real time.

Access requires credentials that match the operating system of the instance. For a Windows instance, Amazon GameLift returns a user name and password as strings for use with a Windows Remote Desktop client. For a Linux instance, Amazon GameLift returns a user name and RSA private key, also as strings, for use with an SSH client. The private key must be saved in the proper format to a .pem file before using. If you're making this request using the AWS CLI, saving the secret can be handled as part of the GetInstanceAccess request. (See the example later in this topic). For more information on remote access, see Remotely Accessing an Instance.

To request access to a specific instance, specify the IDs of the instance and the fleet it belongs to. If successful, an InstanceAccess object is returned containing the instance's IP address and a set of credentials.

fn list_aliases(
&self,
input: &ListAliasesInput
) -> Result<ListAliasesOutput, ListAliasesError>

Retrieves a collection of alias records for this AWS account. You can filter the result set by alias name and/or routing strategy type. Use the pagination parameters to retrieve results in sequential pages.

Aliases are not listed in any particular order.

fn list_builds(
&self,
input: &ListBuildsInput
) -> Result<ListBuildsOutput, ListBuildsError>

Retrieves build records for all builds associated with the AWS account in use. You can limit results to builds that are in a specific status by using the Status parameter. Use the pagination parameters to retrieve results in a set of sequential pages.

Build records are not listed in any particular order.

fn list_fleets(
&self,
input: &ListFleetsInput
) -> Result<ListFleetsOutput, ListFleetsError>

Retrieves a collection of fleet records for this AWS account. You can filter the result set by build ID. Use the pagination parameters to retrieve results in sequential pages.

Fleet records are not listed in any particular order.

fn put_scaling_policy(
&self,
input: &PutScalingPolicyInput
) -> Result<PutScalingPolicyOutput, PutScalingPolicyError>

Creates or updates a scaling policy for a fleet. An active scaling policy prompts Amazon GameLift to track a certain metric for a fleet and automatically change the fleet's capacity in specific circumstances. Each scaling policy contains one rule statement. Fleets can have multiple scaling policies in force simultaneously.

A scaling policy rule statement has the following structure:

If [MetricName] is [ComparisonOperator] [Threshold] for [EvaluationPeriods] minutes, then [ScalingAdjustmentType] to/by [ScalingAdjustment].

For example, this policy: "If the number of idle instances exceeds 20 for more than 15 minutes, then reduce the fleet capacity by 10 instances" could be implemented as the following rule statement:

If [IdleInstances] is [GreaterThanOrEqualToThreshold] [20] for [15] minutes, then [ChangeInCapacity] by [-10].

To create or update a scaling policy, specify a unique combination of name and fleet ID, and set the rule values. All parameters for this action are required. If successful, the policy name is returned. Scaling policies cannot be suspended or made inactive. To stop enforcing a scaling policy, call DeleteScalingPolicy.

fn request_upload_credentials(
&self,
input: &RequestUploadCredentialsInput
) -> Result<RequestUploadCredentialsOutput, RequestUploadCredentialsError>

This API call is not currently in use. Retrieves a fresh set of upload credentials and the assigned Amazon S3 storage location for a specific build. Valid credentials are required to upload your game build files to Amazon S3.

fn resolve_alias(
&self,
input: &ResolveAliasInput
) -> Result<ResolveAliasOutput, ResolveAliasError>

Retrieves the fleet ID that a specified alias is currently pointing to.

fn search_game_sessions(
&self,
input: &SearchGameSessionsInput
) -> Result<SearchGameSessionsOutput, SearchGameSessionsError>

Retrieves a set of game sessions that match a set of search criteria and sorts them in a specified order. Currently a game session search is limited to a single fleet. Search results include only game sessions that are in ACTIVE status. If you need to retrieve game sessions with a status other than active, use DescribeGameSessions. If you need to retrieve the protection policy for each game session, use DescribeGameSessionDetails.

You can search or sort by the following game session attributes:

gameSessionId -- Unique identifier for the game session. You can use either a GameSessionId or GameSessionArn value.

gameSessionName -- Name assigned to a game session. This value is set when requesting a new game session with CreateGameSession or updating with UpdateGameSession. Game session names do not need to be unique to a game session.

creationTimeMillis -- Value indicating when a game session was created. It is expressed in Unix time as milliseconds.

playerSessionCount -- Number of players currently connected to a game session. This value changes rapidly as players join the session or drop out.

maximumSessions -- Maximum number of player sessions allowed for a game session. This value is set when requesting a new game session with CreateGameSession or updating with UpdateGameSession.

hasAvailablePlayerSessions -- Boolean value indicating whether or not a game session has reached its maximum number of players. When searching with this attribute, the search value must be true or false. It is highly recommended that all search requests include this filter attribute to optimize search performance and return only sessions that players can join.

To search or sort, specify either a fleet ID or an alias ID, and provide a search filter expression, a sort expression, or both. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a collection of GameSession objects matching the request is returned.

Returned values for playerSessionCount and hasAvailablePlayerSessions change quickly as players join sessions and others drop out. Results should be considered a snapshot in time. Be sure to refresh search results often, and handle sessions that fill up before a player can join.

Available in Amazon GameLift Local.

fn start_game_session_placement(
&self,
input: &StartGameSessionPlacementInput
) -> Result<StartGameSessionPlacementOutput, StartGameSessionPlacementError>

Places a request for a new game session in a queue (see CreateGameSessionQueue). When processing a placement request, Amazon GameLift searches for available resources on the queue's destinations, scanning each until it finds resources or the placement request times out.

A game session placement request can also request player sessions. When a new game session is successfully created, Amazon GameLift creates a player session for each player included in the request.

When placing a game session, by default Amazon GameLift tries each fleet in the order they are listed in the queue configuration. Ideally, a queue's destinations are listed in preference order.

Alternatively, when requesting a game session with players, you can also provide latency data for each player in relevant regions. Latency data indicates the performance lag a player experiences when connected to a fleet in the region. Amazon GameLift uses latency data to reorder the list of destinations to place the game session in a region with minimal lag. If latency data is provided for multiple players, Amazon GameLift calculates each region's average lag for all players and reorders to get the best game play across all players.

To place a new game session request, specify the following:

The queue name and a set of game session properties and settings

A unique ID (such as a UUID) for the placement. You use this ID to track the status of the placement request

(Optional) A set of IDs and player data for each player you want to join to the new game session

Latency data for all players (if you want to optimize game play for the players)

If successful, a new game session placement is created.

To track the status of a placement request, call DescribeGameSessionPlacement and check the request's status. If the status is Fulfilled, a new game session has been created and a game session ARN and region are referenced. If the placement request times out, you can resubmit the request or retry it with a different queue.

fn stop_game_session_placement(
&self,
input: &StopGameSessionPlacementInput
) -> Result<StopGameSessionPlacementOutput, StopGameSessionPlacementError>

Cancels a game session placement that is in Pending status. To stop a placement, provide the placement ID values. If successful, the placement is moved to Cancelled status.

fn update_alias(
&self,
input: &UpdateAliasInput
) -> Result<UpdateAliasOutput, UpdateAliasError>

Updates properties for a fleet alias. To update properties, specify the alias ID to be updated and provide the information to be changed. To reassign an alias to another fleet, provide an updated routing strategy. If successful, the updated alias record is returned.

fn update_build(
&self,
input: &UpdateBuildInput
) -> Result<UpdateBuildOutput, UpdateBuildError>

Updates metadata in a build record, including the build name and version. To update the metadata, specify the build ID to update and provide the new values. If successful, a build object containing the updated metadata is returned.

fn update_fleet_attributes(
&self,
input: &UpdateFleetAttributesInput
) -> Result<UpdateFleetAttributesOutput, UpdateFleetAttributesError>

Updates fleet properties, including name and description, for a fleet. To update metadata, specify the fleet ID and the property values you want to change. If successful, the fleet ID for the updated fleet is returned.

fn update_fleet_capacity(
&self,
input: &UpdateFleetCapacityInput
) -> Result<UpdateFleetCapacityOutput, UpdateFleetCapacityError>

Updates capacity settings for a fleet. Use this action to specify the number of EC2 instances (hosts) that you want this fleet to contain. Before calling this action, you may want to call DescribeEC2InstanceLimits to get the maximum capacity based on the fleet's EC2 instance type.

If you're using autoscaling (see PutScalingPolicy), you may want to specify a minimum and/or maximum capacity. If you don't provide these, autoscaling can set capacity anywhere between zero and the service limits.

To update fleet capacity, specify the fleet ID and the number of instances you want the fleet to host. If successful, Amazon GameLift starts or terminates instances so that the fleet's active instance count matches the desired instance count. You can view a fleet's current capacity information by calling DescribeFleetCapacity. If the desired instance count is higher than the instance type's limit, the "Limit Exceeded" exception occurs.

fn update_fleet_port_settings(
&self,
input: &UpdateFleetPortSettingsInput
) -> Result<UpdateFleetPortSettingsOutput, UpdateFleetPortSettingsError>

Updates port settings for a fleet. To update settings, specify the fleet ID to be updated and list the permissions you want to update. List the permissions you want to add in InboundPermissionAuthorizations, and permissions you want to remove in InboundPermissionRevocations. Permissions to be removed must match existing fleet permissions. If successful, the fleet ID for the updated fleet is returned.

fn update_game_session(
&self,
input: &UpdateGameSessionInput
) -> Result<UpdateGameSessionOutput, UpdateGameSessionError>

Updates game session properties. This includes the session name, maximum player count, protection policy, which controls whether or not an active game session can be terminated during a scale-down event, and the player session creation policy, which controls whether or not new players can join the session. To update a game session, specify the game session ID and the values you want to change. If successful, an updated GameSession object is returned.

fn update_game_session_queue(
&self,
input: &UpdateGameSessionQueueInput
) -> Result<UpdateGameSessionQueueOutput, UpdateGameSessionQueueError>

Updates settings for a game session queue, which determines how new game session requests in the queue are processed. To update settings, specify the queue name to be updated and provide the new settings. When updating destinations, provide a complete list of destinations.

fn update_runtime_configuration(
&self,
input: &UpdateRuntimeConfigurationInput
) -> Result<UpdateRuntimeConfigurationOutput, UpdateRuntimeConfigurationError>

Updates the current runtime configuration for the specified fleet, which tells Amazon GameLift how to launch server processes on instances in the fleet. You can update a fleet's runtime configuration at any time after the fleet is created; it does not need to be in an ACTIVE status.

To update runtime configuration, specify the fleet ID and provide a RuntimeConfiguration object with the updated collection of server process configurations.

Each instance in a Amazon GameLift fleet checks regularly for an updated runtime configuration and changes how it launches server processes to comply with the latest version. Existing server processes are not affected by the update; they continue to run until they end, while Amazon GameLift simply adds new server processes to fit the current runtime configuration. As a result, the runtime configuration changes are applied gradually as existing processes shut down and new processes are launched in Amazon GameLift's normal process recycling activity.