Struct rusoto_gamelift::CreateFleetInput

pub struct CreateFleetInput {
    pub build_id: Option<String>,
    pub description: Option<String>,
    pub ec2_inbound_permissions: Option<Vec<IpPermission>>,
    pub ec2_instance_type: String,
    pub fleet_type: Option<String>,
    pub instance_role_arn: Option<String>,
    pub log_paths: Option<Vec<String>>,
    pub metric_groups: Option<Vec<String>>,
    pub name: String,
    pub new_game_session_protection_policy: Option<String>,
    pub peer_vpc_aws_account_id: Option<String>,
    pub peer_vpc_id: Option<String>,
    pub resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>,
    pub runtime_configuration: Option<RuntimeConfiguration>,
    pub script_id: Option<String>,
    pub server_launch_parameters: Option<String>,
    pub server_launch_path: Option<String>,
}

Represents the input for a request action.

Fields
build_id: Option<String>

Unique identifier for a build to be deployed on the new fleet. The custom game server build must have been successfully uploaded to Amazon GameLift and be in a READY status. This fleet setting cannot be changed once the fleet is created.

description: Option<String>

Human-readable description of a fleet.

ec2_inbound_permissions: Option<Vec<IpPermission>>

Range of IP addresses and port settings that permit inbound traffic to access game sessions that running on the fleet. For fleets using a custom game build, this parameter is required before game sessions running on the fleet can accept connections. For Realtime Servers fleets, Amazon GameLift automatically sets TCP and UDP ranges for use by the Realtime servers. You can specify multiple permission settings or add more by updating the fleet.

ec2_instance_type: String

Name of an EC2 instance type that is supported in Amazon GameLift. A fleet instance type determines the computing resources of each instance in the fleet, including CPU, memory, storage, and networking capacity. Amazon GameLift supports the following EC2 instance types. See Amazon EC2 Instance Types for detailed descriptions.

fleet_type: Option<String>

Indicates whether to use on-demand instances or spot instances for this fleet. If empty, the default is ON_DEMAND. Both categories of instances use identical hardware and configurations based on the instance type selected for this fleet. Learn more about On-Demand versus Spot Instances.

instance_role_arn: Option<String>

Unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, daemons (background processes). Create a role or look up a role's ARN using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server.

log_paths: Option<Vec<String>>

This parameter is no longer used. Instead, to specify where Amazon GameLift should store log files once a server process shuts down, use the Amazon GameLift server API ProcessReady() and specify one or more directory paths in logParameters. See more information in the Server API Reference.

metric_groups: Option<Vec<String>>

Name of an Amazon CloudWatch metric group to add this fleet to. A metric group aggregates the metrics for all fleets in the group. Specify an existing metric group name, or provide a new name to create a new metric group. A fleet can only be included in one metric group at a time.

name: String

Descriptive label that is associated with a fleet. Fleet names do not need to be unique.

new_game_session_protection_policy: Option<String>

Game session protection policy to apply to all instances in this fleet. If this parameter is not set, instances in this fleet default to no protection. You can change a fleet's protection policy using UpdateFleetAttributes, but this change will only affect sessions created after the policy change. You can also set protection for individual instances using UpdateGameSession.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

peer_vpc_aws_account_id: Option<String>

Unique identifier for the AWS account with the VPC that you want to peer your Amazon GameLift fleet with. You can find your Account ID in the AWS Management Console under account settings.

peer_vpc_id: Option<String>

Unique identifier for a VPC with resources to be accessed by your Amazon GameLift fleet. The VPC must be in the same region where your fleet is deployed. Look up a VPC ID using the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with Amazon GameLift Fleets.

resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>

Policy that limits the number of game sessions an individual player can create over a span of time for this fleet.

runtime_configuration: Option<RuntimeConfiguration>

Instructions for launching server processes on each instance in the fleet. Server processes run either a custom game build executable or a Realtime Servers script. The run-time configuration lists the types of server processes to run on an instance and includes the following configuration settings: the server executable or launch script file, launch parameters, and the number of processes to run concurrently on each instance. A CreateFleet request must include a run-time configuration with at least one server process configuration.

script_id: Option<String>

Unique identifier for a Realtime script to be deployed on the new fleet. The Realtime script must have been successfully uploaded to Amazon GameLift. This fleet setting cannot be changed once the fleet is created.

server_launch_parameters: Option<String>

This parameter is no longer used. Instead, specify server launch parameters in the RuntimeConfiguration parameter. (Requests that specify a server launch path and launch parameters instead of a run-time configuration will continue to work.)

server_launch_path: Option<String>

This parameter is no longer used. Instead, specify a server launch path using the RuntimeConfiguration parameter. (Requests that specify a server launch path and launch parameters instead of a run-time configuration will continue to work.)